+++ src/rules/operation/RuleMintAllowance.sol
@@ -7,7 +7,6 @@ import {RuleInterfaceId} from "RuleEngine/modules/library/RuleInterfaceId.sol";
 import {ERC1404ExtendInterfaceId} from "CMTAT/library/ERC1404ExtendInterfaceId.sol";
 import {RuleEngineInterfaceId} from "CMTAT/library/RuleEngineInterfaceId.sol";
 import {IERC7551Compliance} from "CMTAT/interfaces/tokenization/draft-IERC7551.sol";
-import {IERC3643ComplianceFull} from "../../mocks/IERC3643ComplianceFull.sol";
 import {AccessControlModuleStandalone} from "../../modules/AccessControlModuleStandalone.sol";
 import {RuleMintAllowanceBase} from "./abstract/RuleMintAllowanceBase.sol";
 
@@ -39,11 +38,11 @@ contract RuleMintAllowance is AccessControlModuleStandalone, RuleMintAllowanceBa
         override(AccessControlEnumerable, IERC165)
         returns (bool)
     {
+        // Do not advertise full ERC-3643 ICompliance: its 3-arg mint callback
+        // cannot identify the minter, so quota enforcement needs the spender-aware path.
         return interfaceId == RuleEngineInterfaceId.RULE_ENGINE_INTERFACE_ID
             || interfaceId == ERC1404ExtendInterfaceId.ERC1404EXTEND_INTERFACE_ID
-            || interfaceId == RuleInterfaceId.IRULE_INTERFACE_ID
-            || interfaceId == type(IERC7551Compliance).interfaceId
-            || interfaceId == type(IERC3643ComplianceFull).interfaceId
+            || interfaceId == RuleInterfaceId.IRULE_INTERFACE_ID || interfaceId == type(IERC7551Compliance).interfaceId
             || AccessControlEnumerable.supportsInterface(interfaceId);
     }
 
@@ -55,6 +54,11 @@ contract RuleMintAllowance is AccessControlModuleStandalone, RuleMintAllowanceBa
 
     function _onlyComplianceManager() internal virtual override onlyRole(COMPLIANCE_MANAGER_ROLE) {}
 
-    function _authorizeComplianceBindingChange(address) internal view virtual override onlyRole(COMPLIANCE_MANAGER_ROLE)
+    function _authorizeComplianceBindingChange(address)
+        internal
+        view
+        virtual
+        override
+        onlyRole(COMPLIANCE_MANAGER_ROLE)
     {}
 }